Move global logger setup to pyo3 module
