+++ nautilus_core/Cargo.toml
@@ -26,7 +26,6 @@ chrono = "0.4.26"
 futures = "0.3.28"
 pyo3 = "0.19.2"
 pyo3-asyncio = { version = "0.19.0", features = ["tokio-runtime", "tokio", "attributes"] }
-pyo3-macros = "0.19.2"
 rand = "0.8.5"
 rmp-serde = "1.1.2"
 rust_decimal = "1.31.0"
