+++ nautilus_core/common/Cargo.toml
@@ -19,9 +19,6 @@ serde.workspace = true
 serde_json.workspace = true
 pyo3.workspace = true
 strum.workspace = true
-tracing-appender = "0.2.2"
-tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
-tracing.workspace = true
 ustr.workspace = true
 
 [dev-dependencies]
