+++ nautilus_core/pyo3/Cargo.toml
@@ -11,12 +11,14 @@ name = "nautilus_pyo3"
 crate-type = ["cdylib"]
 
 [dependencies]
-nautilus-common = { path = "../common" }
 nautilus-indicators = { path = "../indicators" }
 nautilus-model = { path = "../model" }
 nautilus-persistence = { path = "../persistence" }
 nautilus-network = { path = "../network" }
 pyo3.workspace = true
+tracing-appender = "0.2.2"
+tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
+tracing.workspace = true
 
 [features]
 extension-module = [
